Test with and without spark (#252)
* Test without spark

* Bump setup python version

* Separate tests out

* Install datacompy step

* Drop 3.0 and add 3.5

* Add fugue no spark

* Add permissions

* Format and docstrings
