+++ tests/test_fugue/test_fugue_helpers.py
@@ -1,0 +1,22 @@
+#
+# Copyright 2020 Capital One Services, LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""One test helper for fugue reports."""
+
+def _compare_report(expected, actual, truncate=False):
+    """Compare datacompy reports."""
+    if truncate:
+        expected = expected.split("Sample Rows", 1)[0]
+        actual = actual.split("Sample Rows", 1)[0]
+    assert expected == actual
